pyproject.toml
[tool.poetry]
name = "smartScore"
version = "0.0.2"
description = ""
authors = ["Nathan Probert <nathanprobert@rogers.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12, <3.13"
requests = {extras = ["socks"], version = "^2.32.3"}
marshmallow = "^3.22.0"
aws-lambda-powertools = "^3.0.0"
python-dotenv = "^1.0.1"
smartscore-info-client = {git = "https://github.com/nathan-probert/smartscore-info-client.git", rev = "v2.2.0"}
pytz = "^2024.2"
unidecode = "^1.3.8"
supabase = "^2.15.0"
beautifulsoup4 = "^4.12.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
ruff = "^0.5.5"
pytest = "^8.2.2"
pytest-cov = "^6.0.0"
pandas = "^2.2.3"
maturin = "^1.8.2"
boto3 = "^1.40.6"
statelint = "^2.0.0"
warp-lang = "^1.11.0"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.9"

[tool.ruff]
exclude = ["**/__pycache__/*", "**/*.pyc", "output/*"]
indent-width = 4
target-version = "py312"
src = ["smartscore", "tests", "templates"]
line-length = 120

[tool.ruff.lint]
select = [
  "BLE",
  "E",
  "F",
  "I",
  "PL",
  "S",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "PLR2004", "S105"]
"smartscore/scripts/*" = ["PLR0915", "PLR0915", "PLR0912", "PLR0913", "E741", "PLR2004"]

[tool.coverage.run]
source = ["smartscore"]
omit = [
    "*/tests/*",
    "*/scripts/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
"tests/*" = [
  "D",
  "S",
  "PLR2004",
  "ARG001",
]
"smartscore/scripts/*" = [
  "PLR0913", # Too many arguments in function definition
  "PLR0912", # Too many branches
  "PLR0915", # Too many statements
  "PLR2004", # Magic value
  "E741",    # Ambiguous variable name
]

[tool.ruff.lint.pylint]
max-args = 8

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
pythonpath = ["smartscore"]
testpaths = ["tests/unit"]

[build-system]
requires = ["maturin>=1.2"]
build-backend = "maturin"
